requirements.txt
pytesseract==0.3.9
requests==2.27.1
SQLAlchemy==1.4.32
Pillow==9.0.1
google-api-python-client==2.41.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1